User guide/BP ref: fix formatting when selecting distro
The PR adding tabs showing the on-prem and hosted versions of BP
customizations and various other options landed somewhat at the same
time as the PR amending documentation to reflect the dot-notation. As a
result, one TOML snipped added by the dot-notation PR was not properly
formatted for the on-prem vs hosted version.

Fix this bug.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>