docs: expand the section on commit messages
This comit expand on the commit messages section and makes it
more clear why they are important and what should be part of them.

There are many good guides on this so it's fine to keep this
brief but I think it was a little bit to concise :)

It also links to a good commit message example, I am open for
suggestions here though, while this is good (as pretty much any
from Christian) I'm sure there are even better ones.