docs/user-guide: add Disk customizations to blueprint reference
Add a description of the disk customizations to the blueprint reference.
Mark is as available on-prem and in bootc-image-builder.

The bootc-image-builder example is in json.  We use both toml and json
in BIB examples and, since the on-prem example is already in toml and it
is a bit of a complicated, deeply-nested structure, having the same
example in json might make it easier to read.