scripts/util.py: enhance resolve_dirs() to handle absolute links
It turned out that absolute links (starting with '/') were not resolved
correctly, which lead to broken links and READMEs not being updated for
more than 2 months.

Extend the resolve_dirs() function to handle this case correctly. The
behavior is that such links to '.md' files are replaces with relative
links within the pulled documentation, while links to any other file
types are converted to full URLs.

Extend the unit tests to cover these scenarios.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>