scripts/utils.py: call resolve_dirs() on .md links starting with /
Although the `resolve_dirs()` have been modified to handle absolute
links for .md files, I forgot to make sure that we call it on such
links in the replace_relative_links() function in the first place.
This is now fixed by adjusting the condition in the function.

In addition, extend the unit tests to explicitly test for this scenario.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>