dev-guide: single-line commits
The guide does not actually elaborate on the fact that multi-line
commits are preferred. Many teams do prefer single-line commits and keep
all the conversation on sites like github.com so it was not obvious for
me.

This change also aligns the capital letter of the first line with what
is the majority of current commits across projects.